+++ src/tests/wallet_test.ts
@@ -4,7 +4,7 @@ import Decimal from "decimal.js";
 import { ethers } from "ethers";
 import { APIError } from "../coinbase/api_error";
 import { Coinbase } from "../coinbase/coinbase";
-import { ArgumentError, InternalError } from "../coinbase/errors";
+import { ArgumentError } from "../coinbase/errors";
 import { Wallet } from "../coinbase/wallet";
 import { Transfer } from "../coinbase/transfer";
 import { ServerSignerStatus, StakeOptionsMode, TransferStatus } from "../coinbase/types";
@@ -43,6 +43,7 @@ import {
   getAssetMock,
   externalAddressApiMock,
   stakeApiMock,
+  walletStakeApiMock,
 } from "./utils";
 import { Trade } from "../coinbase/trade";
 import { WalletAddress } from "../coinbase/address/wallet_address";
@@ -245,6 +246,7 @@ describe("Wallet Class", () => {
 
     beforeAll(() => {
       Coinbase.apiClients.stake = stakeApiMock;
+      Coinbase.apiClients.walletStake = walletStakeApiMock;
       Coinbase.apiClients.asset = assetsApiMock;
     });
 
@@ -258,12 +260,13 @@ describe("Wallet Class", () => {
         STAKING_OPERATION_MODEL.wallet_id = wallet.getId();
         Coinbase.apiClients.asset!.getAsset = getAssetMock();
         Coinbase.apiClients.stake!.getStakingContext = mockReturnValue(STAKING_CONTEXT_MODEL);
-        Coinbase.apiClients.stake!.createStakingOperation =
+        Coinbase.apiClients.walletStake!.createStakingOperation =
           mockReturnValue(STAKING_OPERATION_MODEL);
-        Coinbase.apiClients.stake!.broadcastStakingOperation =
+        Coinbase.apiClients.walletStake!.broadcastStakingOperation =
           mockReturnValue(STAKING_OPERATION_MODEL);
         STAKING_OPERATION_MODEL.status = StakingOperationStatusEnum.Complete;
-        Coinbase.apiClients.stake!.getStakingOperation = mockReturnValue(STAKING_OPERATION_MODEL);
+        Coinbase.apiClients.walletStake!.getStakingOperation =
+          mockReturnValue(STAKING_OPERATION_MODEL);
 
         const op = await wallet.createStake(0.001, Coinbase.assets.Eth);
 
@@ -274,7 +277,7 @@ describe("Wallet Class", () => {
         const newWallet = Wallet.init(walletModel);
         await expect(
           async () => await newWallet.createStake(0.001, Coinbase.assets.Eth),
-        ).rejects.toThrow(InternalError);
+        ).rejects.toThrow(Error);
       });
 
       it("should throw an error when wait is called on wallet address based staking operation", async () => {
@@ -297,12 +300,13 @@ describe("Wallet Class", () => {
         STAKING_OPERATION_MODEL.wallet_id = wallet.getId();
         Coinbase.apiClients.asset!.getAsset = getAssetMock();
         Coinbase.apiClients.stake!.getStakingContext = mockReturnValue(STAKING_CONTEXT_MODEL);
-        Coinbase.apiClients.stake!.createStakingOperation =
+        Coinbase.apiClients.walletStake!.createStakingOperation =
           mockReturnValue(STAKING_OPERATION_MODEL);
-        Coinbase.apiClients.stake!.broadcastStakingOperation =
+        Coinbase.apiClients.walletStake!.broadcastStakingOperation =
           mockReturnValue(STAKING_OPERATION_MODEL);
         STAKING_OPERATION_MODEL.status = StakingOperationStatusEnum.Complete;
-        Coinbase.apiClients.stake!.getStakingOperation = mockReturnValue(STAKING_OPERATION_MODEL);
+        Coinbase.apiClients.walletStake!.getStakingOperation =
+          mockReturnValue(STAKING_OPERATION_MODEL);
 
         const op = await wallet.createUnstake(0.001, Coinbase.assets.Eth);
 
@@ -313,7 +317,7 @@ describe("Wallet Class", () => {
         const newWallet = Wallet.init(walletModel);
         await expect(
           async () => await newWallet.createUnstake(0.001, Coinbase.assets.Eth),
-        ).rejects.toThrow(InternalError);
+        ).rejects.toThrow(Error);
       });
     });
 
@@ -323,12 +327,13 @@ describe("Wallet Class", () => {
         STAKING_OPERATION_MODEL.wallet_id = wallet.getId();
         Coinbase.apiClients.asset!.getAsset = getAssetMock();
         Coinbase.apiClients.stake!.getStakingContext = mockReturnValue(STAKING_CONTEXT_MODEL);
-        Coinbase.apiClients.stake!.createStakingOperation =
+        Coinbase.apiClients.walletStake!.createStakingOperation =
           mockReturnValue(STAKING_OPERATION_MODEL);
-        Coinbase.apiClients.stake!.broadcastStakingOperation =
+        Coinbase.apiClients.walletStake!.broadcastStakingOperation =
           mockReturnValue(STAKING_OPERATION_MODEL);
         STAKING_OPERATION_MODEL.status = StakingOperationStatusEnum.Complete;
-        Coinbase.apiClients.stake!.getStakingOperation = mockReturnValue(STAKING_OPERATION_MODEL);
+        Coinbase.apiClients.walletStake!.getStakingOperation =
+          mockReturnValue(STAKING_OPERATION_MODEL);
 
         const op = await wallet.createClaimStake(0.001, Coinbase.assets.Eth);
 
@@ -339,7 +344,7 @@ describe("Wallet Class", () => {
         const newWallet = Wallet.init(walletModel);
         await expect(
           async () => await newWallet.createClaimStake(0.001, Coinbase.assets.Eth),
-        ).rejects.toThrow(InternalError);
+        ).rejects.toThrow(Error);
       });
     });
 
@@ -348,7 +353,7 @@ describe("Wallet Class", () => {
         const newWallet = Wallet.init(walletModel);
         await expect(
           async () => await newWallet.stakeableBalance(Coinbase.assets.Eth),
-        ).rejects.toThrow(InternalError);
+        ).rejects.toThrow(Error);
       });
 
       it("should return the stakeable balance successfully with default params", async () => {
@@ -364,7 +369,7 @@ describe("Wallet Class", () => {
         const newWallet = Wallet.init(walletModel);
         await expect(
           async () => await newWallet.unstakeableBalance(Coinbase.assets.Eth),
-        ).rejects.toThrow(InternalError);
+        ).rejects.toThrow(Error);
       });
 
       it("should return the unstakeableBalance balance successfully with default params", async () => {
@@ -380,7 +385,7 @@ describe("Wallet Class", () => {
         const newWallet = Wallet.init(walletModel);
         await expect(
           async () => await newWallet.claimableBalance(Coinbase.assets.Eth),
-        ).rejects.toThrow(InternalError);
+        ).rejects.toThrow(Error);
       });
 
       it("should return the claimableBalance balance successfully with default params", async () => {
@@ -396,7 +401,7 @@ describe("Wallet Class", () => {
         const newWallet = Wallet.init(walletModel);
         await expect(
           async () => await newWallet.stakingRewards(Coinbase.assets.Eth),
-        ).rejects.toThrow(InternalError);
+        ).rejects.toThrow(Error);
       });
 
       it("should successfully return staking rewards", async () => {
@@ -413,7 +418,7 @@ describe("Wallet Class", () => {
         const newWallet = Wallet.init(walletModel);
         await expect(
           async () => await newWallet.historicalStakingBalances(Coinbase.assets.Eth),
-        ).rejects.toThrow(InternalError);
+        ).rejects.toThrow(Error);
       });
 
       it("should successfully return historical staking balances", async () => {
@@ -482,7 +487,7 @@ describe("Wallet Class", () => {
           await newWallet.listHistoricalBalances({
             assetId: Coinbase.assets.Usdc,
           }),
-      ).rejects.toThrow(InternalError);
+      ).rejects.toThrow(Error);
     });
 
     it("should successfully return historical balances", async () => {
@@ -687,7 +692,7 @@ describe("Wallet Class", () => {
       expect(wallet.canSign()).toBe(true);
     });
 
-    it("should throw InternalError if derived key is not valid", async () => {
+    it("should throw Error if derived key is not valid", async () => {
       Coinbase.apiClients.address!.listAddresses = mockFn(() => {
         return {
           data: {
@@ -698,7 +703,7 @@ describe("Wallet Class", () => {
           },
         };
       });
-      await expect(wallet.listAddresses()).rejects.toThrow(InternalError);
+      await expect(wallet.listAddresses()).rejects.toThrow(Error);
     });
 
     it("should create new address and update the existing address list", async () => {
@@ -816,7 +821,7 @@ describe("Wallet Class", () => {
         network_id: Coinbase.networks.BaseSepolia,
         feature_set: {} as FeatureSet,
       });
-      await expect(async () => await wallet.faucet()).rejects.toThrow(InternalError);
+      await expect(async () => await wallet.faucet()).rejects.toThrow(Error);
     });
 
     it("should return a Wallet instance", async () => {
@@ -913,7 +918,7 @@ describe("Wallet Class", () => {
 
     it("throws an error when the Wallet is seedless", async () => {
       const seedlessWallet = Wallet.init(walletModel, "");
-      expect(() => seedlessWallet.export()).toThrow(InternalError);
+      expect(() => seedlessWallet.export()).toThrow(Error);
     });
 
     it("should return true for canSign when the wallet is initialized with a seed", () => {
@@ -1112,7 +1117,7 @@ describe("Wallet Class", () => {
 
     it("should throw an error when the wallet is seedless", async () => {
       const seedlessWallet = Wallet.init(walletModel, "");
-      expect(() => seedlessWallet.saveSeed(filePath, false)).toThrow(InternalError);
+      expect(() => seedlessWallet.saveSeed(filePath, false)).toThrow(Error);
     });
   });
 
@@ -1177,7 +1182,7 @@ describe("Wallet Class", () => {
     });
 
     it("raises an error if the wallet is already hydrated", async () => {
-      await expect(seedWallet.loadSeed(filePath)).rejects.toThrow(InternalError);
+      await expect(seedWallet.loadSeed(filePath)).rejects.toThrow(Error);
     });
 
     it("raises an error when file contains different wallet data", async () => {
@@ -1237,7 +1242,7 @@ describe("Wallet Class", () => {
       await expect(
         async () =>
           await newWallet.createTrade({ amount: 0.01, fromAssetId: "eth", toAssetId: "usdc" }),
-      ).rejects.toThrow(InternalError);
+      ).rejects.toThrow(Error);
     });
 
     it("should create a trade from the default address", async () => {